+++ src/contrib/euler/coord.h
@@ -2,9 +2,9 @@ c
 c  Parallel array sizes, including ghost points, for the
 c  mesh coordinates.  Space is allocated in UserSetLocalMesh().
 
-      Double x(cx1:cxn,cy1:cyn,cz1:czn)
-      Double y(cx1:cxn,cy1:cyn,cz1:czn)
-      Double z(cx1:cxn,cy1:cyn,cz1:czn)
+      double precision x(cx1:cxn,cy1:cyn,cz1:czn)
+      double precision y(cx1:cxn,cy1:cyn,cz1:czn)
+      double precision z(cx1:cxn,cy1:cyn,cz1:czn)
 
 c  These parameters enable use of either the local or
 c  global mesh; they are set in the routine parsetup().
