+++ src/contrib/euler/dvarbls.h
@@ -3,11 +3,11 @@ c  Parallel array sizes, including ghost points, for the
 c  residual, corresponding to the F vector in PETSc code.
 c  Space is allocated in UserCreateEuler().
 c
-c      Double  dr(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
-c      Double dru(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
-c      Double drv(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
-c      Double drw(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
-c      Double  de(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
+c      double precision  dr(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
+c      double precision dru(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
+c      double precision drv(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
+c      double precision drw(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
+c      double precision  de(gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
 c
 c  Uniprocessor array sizes:
 c      COMMON /DVRBLS/ DR(NI1,NJ1,NK1),DRU(NI1,NJ1,NK1),DRV(NI1,NJ1,NK1)
@@ -25,5 +25,5 @@ c      COMMON /DVRBLS/ DRW(NI1,NJ1,NK1),DE(NI1,NJ1,NK1)
 #define drw(i,j,k) dxx(4,i,j,k)
 #define de(i,j,k) dxx(5,i,j,k)
 
-        Double dxx(5,gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
+        double precision dxx(5,gxsf1:gxefp1,gysf1:gyefp1,gzsf1:gzefp1)
