+++ src/contrib/euler/init.h
@@ -2,5 +2,5 @@ c
 c  Initial conditions
 c
       COMMON /INIT/ RINIT,RUINIT,RVINIT,RWINIT,EINIT
-      Double RINIT,RUINIT,RVINIT,RWINIT,EINIT
+      double precision  RINIT,RUINIT,RVINIT,RWINIT,EINIT
 c
